tools/ai_search/requirements.txt
sentence-transformers==3.1.1
supabase==2.9.1
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
